Play a game of coup with LLM players
Cards are handed out at random. Static models are chosen and each loaded
with different personalities. With enough runs, perhaps we can garner
some understanding of how different characteristics for LLMs perform in
play. In part, it can be a social modeling experiment. A large part of
the motivation is to see how LLMs perform in situations that reward
deception and strategizing.

We use a typescript basic CLI, a model-agnostic, OpenAI-API compatible
calling interface with function calling. At each turn each player can
decide a variety of actions. Possessing a particular card that
explicitly gives them powers to take that action is reliable, safe bet.
However, they are also able to bluff their way entirely throughout the
game and take actions which they do not have explicit permission for.